feat(feeders): long() honors full Long range + generators test suite
- long(min,max) gated with Long.MaxValue/MinValue (was Int.*), so bounds
  outside the Int range no longer fall through to an unbounded draw (#205)
- ** / repeat / separateBy reject n <= 0 with a clear IllegalArgumentException
  instead of UnsupportedOperationException: empty.reduceLeft (#205)
- add GeneratorsSpec: first test coverage for the feeders/generators package
  (numeric/string/char/collection generators, ~/**/separateBy combinators,
  case-class derivation, seeded determinism)

+++ src/main/scala/org/galaxio/gatling/feeders/generators/GeneratorInstances.scala
@@ -23,8 +23,8 @@ trait GeneratorInstances {
 
   final def long(min: Long, max: Long): Generator[Long] = Generator { ctx =>
     val res =
-      if (max < Int.MaxValue) ctx.gen.between(min, max + 1)
-      else if (min > Int.MinValue) ctx.gen.between(min - 1, max) + 1
+      if (max < Long.MaxValue) ctx.gen.between(min, max + 1)
+      else if (min > Long.MinValue) ctx.gen.between(min - 1, max) + 1
       else ctx.gen.nextLong()
 
     res
